# backend/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Backend
host: localhost:4000
...
basePath: /api
schemes:
  - http
  - https
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
     For accessing the API a valid JWT token must be passed in all the queries in
     the 'Authorization' header.
  
  
     A valid JWT token is generated by the API and retourned as answer of a call
     to the route /login giving a valid user & password.
  
  
     The following syntax must be used in the 'Authorization' header :
  
         Bearer xxxxxx.yyyyyyy.zzzzzz
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /groups/{idGroup}/subgroups/{idSubgroup}/questions:
    x-swagger-router-controller: questions.controller
    parameters:
      - name: idGroup
        in: path
        description: "Id of the Group"
        required: true
        type: string
      - name: idSubgroup
        in: path
        description: "Id of the Subgroup"
        required: true
        type: string
    post:
      tags:
        - Questions
      operationId: questionsControllerPost
      summary: "Create Question"
      description: "Create a new Question"
      parameters:
        - name: Question
          in: body
          required: true
          schema:
            $ref: '#/definitions/Question'
      responses:
        "200":
          description: "Success"
        "403":
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
    get:
      tags:
        - Questions
      operationId: questionsControllerGet
      summary: "Get Question"
      description: "Get all the Question"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: '#/definitions/Question'
        "403":
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
  /groups/{idGroup}/subgroups/{idSubgroup}/questions/{idQuestion}:
    x-swagger-router-controller: questions.controller
    parameters:
      - name: idGroup
        in: path
        description: "Id of the Group"
        required: true
        type: string
      - name: idSubgroup
        in: path
        description: "Id of the Subgroup"
        required: true
        type: string
      - name: idQuestion
        in: path
        description: "ID of the Question"
        required: true
        type: string
    get:
      tags:
        - Questions
      operationId: questionsControllerGetId
      summary: "Get Question"
      description: "Get only one Question"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: '#/definitions/Question'
        "403":
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'

    put:
      tags:
        - Questions
      operationId: questionsControllerPut
      summary: "Update a Question"
      description: "Update only one Question"
      parameters:
        - name: Question
          in: body
          required: true
          schema:
            $ref: '#/definitions/Question'
      responses:
        200:
          description: "Success"
          
    delete:
      tags:
        - Questions
      operationId: questionsControllerDelete
      summary: "Delete a Question"
      description: "Delete only one Question"
      responses:
        200:
          description: "Success"
...
  /detailProcedure/{idDetailTypeProcedure}/procedure:
    x-swagger-router-controller: procedures.controller
    parameters:
      - name: idDetailTypeProcedure
        in: path
        description: "ID of Detail od Type of Procedure"
        required: true
        type: string
    post:
      tags:
        - Procedure
      operationId: proceduresControllerPost
      summary: "Create Procedure"
      description: "Create a new Procedure"
      parameters:
        - name: Procedure
          in: body
          required: true
          schema:
            $ref: '#/definitions/Procedure'
      responses:
        200:
          description: "Success"
        403:
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
    get:
      tags:
        - Procedure
      operationId: proceduresControllerGet
      summary: "Get Procedure"
      description: "Get all the Procedure"
      responses:
        200:
          description: "Success"
          schema:
            $ref: '#/definitions/Procedure'
        403:
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
  /detailProcedure/{idDetailTypeProcedure}/procedure/{idProcedure}:
    x-swagger-router-controller: procedures.controller
    parameters:
      - name: idProcedure
        in: path
        description: "ID of the Person"
        required: true
        type: string
      - name: idDetailTypeProcedure
        in: path
        description: "ID of Detail od Type of Procedure"
        required: true
        type: string
    get:
      tags:
        - Procedure
      operationId: proceduresControllerGetId
      summary: "Get procedure"
      description: "Get only one Person"
      responses:
        200:
          description: "Success"
          schema:
            $ref: '#/definitions/Person'
        403:
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'

    put:
      tags:
        - Procedure
      operationId: proceduresControllerPut
      summary: "Update a procedure"
      description: "Update only one procedure"
      parameters:
        - name: Procedure
          in: body
          required: true
          schema:
            $ref: '#/definitions/Procedure'
      responses:
        200:
          description: "Success"
        403:
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
          
    delete:
      tags:
        - Procedure
      operationId: proceduresControllerDelete
      summary: "Delete a procedure"
      description: "Delete only one procedure"
      responses:
        200:
          description: "Success"
        403:
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  Procedure:
    type: object
    properties:
      id:
        type: string
      idTypeOfProcedure:
        type: string
        example: "497024ee-2542-11ea-9214-ac220b26bfde"
      idDetailProcedure:
        type: string
        example: "6b0749c7-25d7-11ea-9214-ac220b26bfde"
      date:
        type: string
        example: "09/09/2019"
      institutionId:
        type: string
        example: "e034cefc-25d7-11ea-9214-ac220b26bfde"
      serviceId:
        type: string
        example: "dbb25569-25d7-11ea-9214-ac220b26bfde"
      patientId:
        type: string
        example: "d7246687-25d7-11ea-9214-ac220b26bfde"
      description:
        type: string
        example: "Operacion de rodilla realizada al senior bla bla en el sanatorio allende centro a cargo del servicio de cirugia."
...
  Question:
    type: object
    properties:
      id:
        type: string
      question:
        type: string
        example: "Subiendo y bajando escaleras"
  Person:
    type: object
    properties:
      id:
        type: string
      dni:
        type: string
        example: "37195387"
      birthdate:
        type: string
        example: "09/01/1993"
      firstname:
        type: string
        example: "Federico Martin"
      lastname:
        type: string
        example: "Cabanilla"
...
  Error:
    type: object
    properties:
      message:
        type: string